fix(pagamento): convert valor to long integer when calling procpag API
Change RequisicaoProcpagDTO.valor from BigDecimal to Long to match the
procpag OpenAPI spec (type: number, format: int). Pass valorTotal.longValue()
when constructing the DTO in ProcessarPagamentoUseCase.

+++ pagamento-service/src/main/java/com/fiap/pagamento/core/dto/RequisicaoProcpagDTO.java
@@ -6,14 +6,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class RequisicaoProcpagDTO {
-    private BigDecimal valor;
+    private Long valor;
 
     @JsonProperty("pagamento_id")
     private String pagamentoId;
